+++ lib/database/migrations/all_migrations.dart
@@ -22,6 +22,7 @@ import 'package:pantry_app/database/migrations/v28_normalize_produce_barcodes.da
 import 'package:pantry_app/database/migrations/v29_inventory_unique_index.dart';
 import 'package:pantry_app/database/migrations/v2_inventories.dart';
 import 'package:pantry_app/database/migrations/v30_recipe_indexes_and_search.dart';
+import 'package:pantry_app/database/migrations/v31_serving_quantity.dart';
 import 'package:pantry_app/database/migrations/v3_normalize_units.dart';
 import 'package:pantry_app/database/migrations/v4_nutriscore_grade.dart';
 import 'package:pantry_app/database/migrations/v5_nutriscore_not_applicable.dart';
@@ -62,4 +63,5 @@ List<Migration> allMigrations() => [
   MigrationV28(),
   MigrationV29(),
   MigrationV30(),
+  MigrationV31(),
 ];